% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpColumnPlotsR.R
\name{dpColumnPlotsR}
\alias{dpColumnPlotsR}
\title{Data Profile Column Plot Data via R}
\usage{
dpColumnPlotsR(in_project_id, in_table_id, input_tbl, dp_col_prop,
  max_levels = 12)
}
\arguments{
\item{in_project_id}{character The UUID ID for the project in
which this table is to be included.}

\item{in_table_id}{integer The table sequence number for table to analize.}

\item{input_tbl}{data.frame The data frame to analize.}

\item{dp_col_prop}{data.frame The data frame with the column properties of
\code{input_tbl} that has been previously generated.}

\item{max_levels}{= 12 The maximum number of levals to include in a catigorical plot}
}
\value{
data.frame with a row for each column in input_tbl and columns:
\itemize{
  \item project_id - UUID of project with this table
  \item table_id - of table with these columns
  \item column_id - integer sequence of column
  \item plot_id - intially = 1; may eventually have more than one plot per column
  \item plot_type - one of ["catigorical", "first character", "histogram"]
  \item x_var character - label of the bar
  \item count integer - number of rows in bar's value or range
}
}
\description{
\code{dpColumnPlotsR} calculates the values for column level plots.
}
\details{
This design separates the plot data generation for plotting. There are two
advantages of this design. First, for large data we want to offload the calculation
to the data platform so we only get back reduced data. Secondly, different report
designs will require unique plot aesthetics.

Generates the data for a bar or histogram plot. Note \code{x_var} is a character and
should be converted to numeric when ploting a histogram.

For DBMS resident data you should use \code{dpColumnPlotsR} to avoid downloading all
rows to R.
}
\examples{
\dontrun{
dpProjID <- dpTables$project_id[1]
dpTblID <- 1
Tbl <- read_csv(dpTables$table_source[dpTblID])
dpColProp <- dpColumnPropertiesR(dpProjID, dpTblID, Tbl)
dpColPlts <- dpColumnPlotsR(dpProjID, dpTblID, Tbl, dpColProp)
}
}

