% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpColumnPropertiesR.R
\name{dpColumnPropertiesR}
\alias{dpColumnPropertiesR}
\title{Data Profile Column Properties via R}
\usage{
dpColumnPropertiesR(in_project_id, in_table_id, input_tbl,
  collapse_with = "|")
}
\arguments{
\item{in_project_id}{character The UUID ID for the project in
which this table is to be included.}

\item{in_table_id}{integer The table sequence number for table to analize.}

\item{input_tbl}{data.frame The data frame to analize.}

\item{collapse_with}{= "|" character The characer string used with \code{collapse}}
}
\value{
data.frame with a row for each column in input_tbl and columns:
\itemize{
  \item project_id - UUID of project with this table
  \item table_id - of table with these columns
  \item column_id - integer sequence of column
  \item column_name - from table meta-data
  \item column_type - from table meta-data. NULL if no meta-data in which case
    all columns loaded as text
  \item column_content_type - most predominate type discovered in text date. NULL
    not text.
  \item column_dommain - as discovered by scan. PLACE HOLDER FOR VER 0.2 (URL,
    ZIP, Phone, Date, Age, ...)
  \item num_rows - number of rows in table (nrow(input_tbl))
  \item num_nulls - number NA's in column
  \item num_distinct - number distinct values in column
  \item num_empty - if character, number null-strings (as distinct from NAs), else NA
  \item num_numeric - if character, number represinting numbers, else NA
  \item num_date - if character, number representing dates or timestamps, else NA
  \item mean - NA if not numeric
  \item minimum - smallest value as string
  \item first_quartile - value at 25% along sorted vector as string
  \item second_quartile - value at 50% along sorted vector as string
  \item third_quartile - value at 75% along sorted vector as string
  \item maximum - largest value as string
  \item min_length - minimum number of characters or significant digits
  \item max_lenght - maximum number of characters or significant digits
  \item head - collapsed \code{head()} limited to 250 characters
  \item tail - collapsed \code{tail()} limited to 250 characters
  \item bottom - collapsed \code{head(sort())} limited to 250 characters
  \item top - collapsed \code{tail(sort())} limited to 250 characters
}
}
\description{
\code{dpColumnPropertiesR} calculates the column properties of an R data frame.
}
\details{
The data frame is typically loaded from a text file, Excel sheet, or other tabular
structure.

For DBMS resident data you should use dpColumnPropertiesSQL to avoid downloading
all rows to R.
}
\examples{
\dontrun{
dpProjID <- dpTables$project_id[1]
dpTblID <- 1
Tbl <- read_csv(dpTables$table_source[dpTblID])
dpColProp <- dpColumnPropertiesR(dpProjID, dpTblID, Tbl)
}
}

