% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpMakeProject.R
\name{dpMakeProject}
\alias{dpMakeProject}
\title{Make a project data frame.}
\usage{
dpMakeProject(name, description, created_by, notes = NA)
}
\arguments{
\item{name}{character Short name of the project.}

\item{description}{character Longer description of project.}

\item{created_by}{character Person or system creating project.}

\item{notes=NA}{character Whatever notes needed for project.}
}
\value{
data.frame Initialized of one row with columns:
\itemize{
  \item project_id - UUID generated for this row
  \item project_name - from argument list
  \item project_description - from argument list
  \item project_notes - optional, from argument list
  \item created_by - from argument list
  \item created_at - Sys.time()
  \item mod_by - initially NA
  \item mod_at - initially NA
}
}
\description{
\code{dpMakeProject} creates a initial one-row data frame for
the data profiling project.
}
\examples{
\dontrun{
dpMakeProject("Test1", "A test project", "JimP")}
}

