% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpMakeTable.R
\name{dpMakeTable}
\alias{dpMakeTable}
\title{Make a table data frame.}
\usage{
dpMakeTable(in_project_id, name, source, description, created_by, notes = NA)
}
\arguments{
\item{in_project_id}{character The UUID ID for the project in
which this table is to be included.}

\item{name}{character R data.frame or DBMS table name}

\item{source}{character The literal string to connect to the table. Will
be a function of environment; eg text file, DBMS, R, etc.}

\item{description}{character A longer description of table.}

\item{created_by}{character Person or system creating this table entry.}

\item{notes}{character Notes about this table.}
}
\value{
data.frame Initialized to one row with columns:
\itemize{
  \item project_id - UUID of project with this table
  \item table_id - integer sequence of tables within project
  \item table_name - from argument list
  \item table_source - from argument list
  \item table_description
  \item table_notes- from argument list
  \item table_rows - # rows in table
  \item table_columns - # columns in table
  \item created_by - from argument list
  \item created_at - Sys.time()
  \item mod_by - initially NA
  \item mod_at - initially NA
}
}
\description{
\code{dpMakeTable} creates an initial one-row data frame for
a table in the data profiling project.
}
\examples{
\dontrun{
TblName <- "NewsCurrentEvents"
TblSource <- system.file("extdata", "NewsCurrentEvents.csv",
                          package = "dProf")
dpTables <- dpMakeTable(dpTest$project_id[1], TblName, TblSource, "Jim P",
                        notes = "Using the example csv in project")
str(dpTables)
}
}

