% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiveval.R
\name{fiveval}
\alias{fiveval}
\title{Calculate the five values of a vector.}
\usage{
fiveval(x)
}
\arguments{
\item{x}{vector On which the five values will be calculated.}
}
\value{
a 5-element character vector with elements named: Min, 1st Qtl,
  2nd Qtl, 3rd Qtl, and Max.
}
\description{
\code{fiveval} is a generalization of \code{fivenum} to allow non-numeric input
}
\details{
Unlike \code{fivenum}, \code{fiveval} returns a character vector and will
accept any data type. For non-numerics, each value is just picked out of the
ordered list of values using the \code{fivenum} indexing scheme. Instead of "median", we use the "2nd Qtl" to make it
clear this is not \code{fivenum}.
}
\examples{
  fiveval(-100:100)
  fiveval(LETTERS)
  fiveval(seq(as.Date("1990-01-01"), as.Date("2009-12-31"), by = 1))
}

