% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try.Date.R
\name{try.Date}
\alias{try.Date}
\title{Look for dates in a character vector.}
\usage{
try.Date(try_this)
}
\arguments{
\item{try_this}{character Vector to test for Date compatible elements}
}
\value{
integer maximum number of Date compatible elements found
}
\description{
\code{try.Date} tries common date formats on elements of a character vector
and returns the maximum number of dates found.
}
\details{
Implicit assumption is Dates in vector have same format.
}
\examples{
try.Date(c("1/1/1999", "12/1/2020", "abcd"))
try.Date(c("12/30/1999", "30/12/1999", "1999-12-30"))
}

